Figured out what to do. Almost done with the project!
This project seems to be pretty easy. We are implementing a very
simple version of a direct-mapped cache. The code is pretty short
so far, so analyzing it will probably be enough to understand
what is happening. I will go through and comment things and
actually finish the code soon.
Had an issue where I used a wrong mask for the index value, which
made us fail most of the tests online.
Old mask: & calcLog(inCache->cacheSize);
New mask: & inCache->cacheSize - 1;
We'll see if it works!